Moved source files under lib/
