+++ lib/components/ControlPanel.jsx
