+++ lib/components/FileLegend.jsx
