+++ lib/components/MemoryLayout.jsx
