+++ lib/containers/controlPanel.js
