+++ lib/hexpad.js
