+++ lib/memRegions.js
