+++ lib/reducers/fileReducer.js
