+++ lib/reducers/index.js
