+++ lib/reducers/targetReducer.js
